using singature builder in rest client; refactoring
