+++ .gitignore
@@ -11,10 +11,10 @@ unity/Tests/ScoreboardService.Tests/obj/
 
 # Autogenerated VS/MD solution and project files
 ExportedObj/
-*.csproj
+unity/*.csproj
+unity/*.sln
+unity/*.suo
 *.unityproj
-*.sln
-*.suo
 *.tmp
 *.user
 *.userprefs
@@ -30,8 +30,8 @@ ExportedObj/
 sysinfo.txt
 
 # Builds
-*.apk
-*.unitypackage
+unity/*.unitypackage
+unity/*.apk
 
 src/UnityClient/bin/
 src/UnityClient/obj/
